refactor: (#531) manager api student domain으로 이동
