+++ dms-core/src/main/kotlin/team/aliens/dms/domain/student/usecase/QueryStudentDetailsUseCase.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.domain.manager.usecase
+package team.aliens.dms.domain.student.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
 import team.aliens.dms.domain.point.service.PointService
