+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/global/security/SecurityConfig.kt
@@ -58,17 +58,17 @@ class SecurityConfig(
             .antMatchers(HttpMethod.PATCH, "/students/password/initialization").permitAll()
             .antMatchers(HttpMethod.PATCH, "/students/profile").hasAuthority(STUDENT.name)
             .antMatchers(HttpMethod.DELETE, "/students").hasAuthority(STUDENT.name)
+            .antMatchers(HttpMethod.GET, "/students").hasAuthority(MANAGER.name)
+            .antMatchers(HttpMethod.GET, "/students/{student-id}").hasAuthority(MANAGER.name)
+            .antMatchers(HttpMethod.DELETE, "/students/{student-id}").hasAuthority(MANAGER.name)
+            .antMatchers(HttpMethod.GET, "/students/file").hasAuthority(MANAGER.name)
+            .antMatchers(HttpMethod.POST, "/students/file/room").hasAuthority(MANAGER.name)
+            .antMatchers(HttpMethod.POST, "/students/file/gcn").hasAuthority(MANAGER.name)
 
             // /managers
             .antMatchers(HttpMethod.GET, "/managers/account-id/{school-id}").permitAll()
-            .antMatchers(HttpMethod.GET, "/managers/students").hasAuthority(MANAGER.name)
-            .antMatchers(HttpMethod.GET, "/managers/students/{student-id}").hasAuthority(MANAGER.name)
             .antMatchers(HttpMethod.GET, "/managers/profile").hasAuthority(MANAGER.name)
             .antMatchers(HttpMethod.PATCH, "managers/password/initialization").permitAll()
-            .antMatchers(HttpMethod.DELETE, "/managers/students/{student-id}").hasAuthority(MANAGER.name)
-            .antMatchers(HttpMethod.GET, "/managers/students/file").hasAuthority(MANAGER.name)
-            .antMatchers(HttpMethod.POST, "/managers/students/file/room").hasAuthority(MANAGER.name)
-            .antMatchers(HttpMethod.POST, "/managers/students/file/gcn").hasAuthority(MANAGER.name)
 
             // /schools
             .antMatchers(HttpMethod.GET, "/schools").permitAll()
